feat(UI): implement HomePage and move there logic from App
