+++ src/pages/HomePage/index.tsx
@@ -1,0 +1,1 @@
+export { HomePage } from './HomePage'
